+++ src/app/gallery/page.tsx
@@ -1,13 +1,23 @@
 import Title from "@/components/Title";
-import Events from "@/components/gallery/Event";
 import Gallery from "@/components/gallery/Gallery";
-
+import Image from "next/image";
+import pinkFlower from "@/public/gallery/pinkFlower.webp";
+import blueFlower from "@/public/gallery/blueFlower.webp";
 const Page = () => {
   return (
-    <div className="bg-csa-tan-100">
+    <div className="relative min-h-screen bg-csa-tan-100">
       <Title>gallery</Title>
-      <Events />
       <Gallery />
+      <Image
+        src={pinkFlower}
+        alt="pinkFlower"
+        className="absolute bottom-0 left-0 w-4/12 md:w-2/12"
+      />
+      <Image
+        src={blueFlower}
+        alt="blueFlower"
+        className="absolute bottom-0 right-0 w-4/12 md:w-2/12"
+      />
     </div>
   );
 };
